fix: persist focus when switching tabs
For the callbacks related to persisting focus when switching tabs, switched from a timeout based approach to cancel the callbacks to a method that looks for change in focus. The timeout based method was not working on Chrome and Firefox in Windows. The new approach still ensures that only the most recent math-field element will have a `visibilitychange` callback associated with it.